Suma de cantidades de domicilios en reporte de domicilios día

+++ netbeans/sede/src/main/java/com/administracion/service/jsf/DomiciliosDiaColombianService.java
@@ -15,6 +15,6 @@ import java.util.List;
 public interface DomiciliosDiaColombianService {
     
     public List<OrdenesColombianDto> domicilioDia(Date fi, Date ff,String subsede);
-    public Double getTotalDomicilios();
+    public Long getTotalDomicilios();
     public Long getTotalRegistros();
 }
